week-6/advanced/code-review-app/requirements.txt
streamlit
openai
langchain
python-dotenv
chromadb
huggingface_hub
sentence-transformers